Reduce the try_receive timeout to 10ms
While we're waiting in try_receive, we're not sending out pending TDs
that might be ready. Reduce the timeout a bit so that we don't delay
those ready TDs too long.

A better solution would be to work out when the next TD will be ready
and only sleep till then, but that's too much effort for now